Prestige Lavender Fields
Prestige Lavender Fields Whitefield Bangalore is a high-end residential project that has developed recently in the IT City of India. This is a nature theme project which will advancement your lifestyle and connect you with nature. The project has been deliberate in-house by Prestige Group.
Stay at the highest point of Bangalore where nature and grandeur will welcome you. Coming up with the finest residential project that promises to give a taste of the natural city life. This flat floor plan offers 1, 2 , 3 and 4 BHK luxury homes.
Each apartment is imbued with the luxury of world-class amenities. This new property in Bangalore is spread across sprawling acres of land with abundant outdoor spaces. Development is 80% of the total extent of open spaces.
You will love the open spaces with greenery all around at Prestige Lavender Fields Bangalore. Wide balconies and windows will permit you enough ventilation and light.
Guarantees you outstanding profits and growth in the upcoming future. It will provide you the most luxurious life between nature and city life.
The new residential project is located in one of the most desirable localities of South Bangalore. The location upholds proximity to various areas of the city. The presence of NICE Expressway further enhances the connectivity of the area. This project review is amazing as many home buyers are eagerly waiting for this project.
Whitefield Bangalore has developed a popular real estate destination for home buyers and investors since the expansion of metro connectivity.
Prestige Lavender Fields is the best residence in Whitefield Bangalore, which minimizes wasted space and offers excellent utilization of space. The spacious layout, systematic planning, and world-class amenities all combine to make this a paradisiacal home.

For more information - 
Visit: https://bit.ly/3Yg29Zr